Move sizes enums into the race module
